[IMP] dms: Show all directories in searchpanel from files
TT46388

[FIX] dms: Set the correct parent_id of directories in searchpanel

If the parent directory is not in all the records we should not set
parent_id because the user does not have access to parent.

TT46388

[FIX] dms: Avoid error when filtering in files

If we filter on files (for example by name) we should not apply kwargs search_domain
because we would be incorrectly filtering the directories by name, and we should
always list them all.

TT46903

(cherry picked from commit 825e5725d157e5022e18d4e280d4e25ca4c6d5e7)