[FIX] dms: Delete the file if the attachment is deleted
When deleting an attachment, an attempt should be made to delete the linked file, causing an error if it is locked.
Do not show the Delete/Archive/Unarchive option on files if it is locked and you cannot unlock it.

Fixes https://github.com/OCA/dms/issues/381

(cherry picked from commit 60d11587db00af22656300884c056577199df4d9)